updated android toolchain to ndk-r8e with more, but not complete c++11 features
- thread and mutex library is now available
- std::string::to_string() still missing
