--- android/WiflyLight/jni/c++11_wrapper/atomic
@@ -1,42 +1,0 @@
-/**
- Copyright (C) 2013 Nils Weiss, Patrick Bruenn.
- 
- This file is part of Wifly_Light.
- 
- Wifly_Light is free software: you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation, either version 3 of the License, or
- (at your option) any later version.
- 
- Wifly_Light is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
- 
- You should have received a copy of the GNU General Public License
- along with Wifly_Light.  If not, see <http://www.gnu.org/licenses/>. */
-
-#if defined(__cplusplus) && (__cplusplus < 201103L)
-#warning "Check for a newer compiler to avoid using this C++11 wrapper file"
-#ifndef _CPP11_WRAPPER_ATOMIC_
-#define _CPP11_WRAPPER_ATOMIC_
-namespace std {
-template<class T>
-struct atomic {
-	T data;
-	atomic(T v) : data(v) {};
-};
-
-template<typename T>
-inline T atomic_fetch_add(atomic<T>* pObj, T value) {
-	return __sync_fetch_and_add(&pObj->data, value);
-};
-
-template<typename T>
-inline T atomic_fetch_sub(atomic<T>* pObj, T value) {
-	return __sync_fetch_and_sub(&pObj->data, value);
-};
-} /* namespace std { */
-#endif /* #ifndef _CPP11_WRAPPER_ATOMIC_ */
-#endif /* #if defined(__cplusplus) && (__cplusplus < 201103L) */
-
