+++ library/FwResponse.h
@@ -27,6 +27,19 @@
 #include "wifly_cmd.h"
 #include "WiflyControlException.h"
 
+#ifdef ANDROID
+#warning "Check for a new android-ndk > r8e to avoid using this selfmade implementation"
+namespace std {
+	template <class T>
+	std::string to_string(T value)
+	{
+		std::stringstream converter;
+		converter << value;
+		return converter.str();
+	}
+}
+#endif /* #ifdef ANDROID */
+
 namespace WyLight {
 
 class FwResponse
